Show all edges between two vertices
This ensures that an edge AB will not overlap with another edge AB,
where A and B are connected vertices.

The "dynamic" edge type automatically avoids other edges with the same
vertices where the previously used "cubicBezier" type does not. See:

    https://github.com/almende/vis/issues/1957

Unfortunately, this loses the cubic curve in favour of a quadratic one.